requirements.txt
tensorflow>=2.12.0
streamlit>=1.24.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.1
h5py>=3.8.0
